chanstate: move forwarding kv codecs
Move AddRef and PkgFilter persistence codecs from the forwarding domain
file into the forwarding KV store file. The forwarding domain types keep
their pure behavior while the KV-specific byte encoding stays next to
the bucket code that stores those values.
